Move Balances to component
